pyproject.toml
[tool.poetry]
name = "treker"
version = "0.1.0"
description = ""
authors = ["Elizaveta <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
requests = "^2.32.4"
python-dotenv = "^1.1.1"
celery = "^5.5.3"
djangorestframework-simplejwt = "^5.5.0"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
